+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.1-beta.3](https://pypi.org/project/Pyntel4004/0.0.1b3/) - 2021-XX-XX
+
+### Notes
+-
+### Added
+- Additonal opcode documentation
+### Changed
+-
+### Deprecated
+-
+### Fixed
+-
+### Removed
+-
+### Security
+-
+
 ## [0.0.1-beta.2](https://pypi.org/project/Pyntel4004/0.0.1b2/) - 2021-12-11
 
 ### Notes